refactor: move components and theme into UI feature
