+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../theme/globals.css'
+import '~/features/ui/theme/globals.css'
 
 // @ts-ignore // Types to be added in TS lesson
 function MyApp({ Component, pageProps }) {
